Message the user's confirmed email
Before, someone could put any email in when filling out the Developer
form. Now, the only way to change that is to change your account's email
address.

In the future, this might change to allow enabling new emails. But for
now this works well enough.

developers.email is now deprecated, but the data remains in case we need
to bring it back in the future.