Replace custom block column handling with dedicated model
Before we were doing handling of blocks by swapping between columns based
on who's a developer and who's a business, but we can obviate the need to
know that by tracking the change as a new association between participants.

This is a step in the direction of a conversation not needing to know that
either participant is a business or a developer.