Add Read At Indicator
This PR should add a new checkmark read at indicator if the message has
been read and is the last message in the conversation. The popup does
not look the best but can adjust with any feedback. There is a case in
the fixtures where a message can have two notifications, is that a use
case to worry about?

* Add a new feature `message_read_indicator`
* Add a new component `Messages::ReadAt`