Multiple Tailwind CSS configs (for the redesign) (#701)
* Scaffolding for switching between Tailwind configs

Note that the `redesign` yarn script might not get run in production. I
_think_ that Heroku automatically picks up `build:css` so we might need
a way to tell Heroku to run the new one, too.

* New design colors!

Note how the theme was not extended but "replaced". This means that
`text-orange-100` will have no effect because the `orange` color is not
defined.

This will serve as a halfway point until we can recognize which colors
we need and are less brand-related and more contextual (errors and
such).

* Enable redesign with `REDESIGN=1 bin/dev`

Or by setting the `REDESIGN` environment variable in production.