Move developer response rate into Developer model.
The developer response rate is stored in the `developers` table. This
value will be updated a configurable time after a new conversation is
intitiated by a Business.
This commit includes a migration to add the db column.
This commit includes a rake task (rake
backfills:developer_response_rate) to be run after the migration.